requirements.txt
yt-dlp
gspread
oauth2client
google-api-python-client
python-dotenv
faster-whisper